feat: add utility GetBackupStoreDriver
