+++ pkg/util/backup.go
@@ -1,0 +1,44 @@
+package util
+
+import (
+	"fmt"
+	"os"
+	"strings"
+
+	"github.com/longhorn/backupstore"
+	ctlcorev1 "github.com/rancher/wrangler/pkg/generated/controllers/core/v1"
+
+	"github.com/harvester/harvester/pkg/settings"
+)
+
+func ConstructEndpoint(target *settings.BackupTarget) string {
+	switch target.Type {
+	case settings.S3BackupType:
+		return fmt.Sprintf("s3://%s@%s/", target.BucketName, target.BucketRegion)
+	case settings.NFSBackupType:
+		// we allow users to input nfs:// prefix as optional
+		return fmt.Sprintf("nfs://%s", strings.TrimPrefix(target.Endpoint, "nfs://"))
+	default:
+		return target.Endpoint
+	}
+}
+
+func GetBackupStoreDriver(secretCache ctlcorev1.SecretCache, target *settings.BackupTarget) (backupstore.BackupStoreDriver, error) {
+	if target.Type == settings.S3BackupType {
+		secret, err := secretCache.Get(LonghornSystemNamespaceName, BackupTargetSecretName)
+		if err != nil {
+			return nil, err
+		}
+		os.Setenv(AWSAccessKey, string(secret.Data[AWSAccessKey]))
+		os.Setenv(AWSSecretKey, string(secret.Data[AWSSecretKey]))
+		os.Setenv(AWSEndpoints, string(secret.Data[AWSEndpoints]))
+		os.Setenv(AWSCERT, string(secret.Data[AWSCERT]))
+	}
+
+	endpoint := ConstructEndpoint(target)
+	bsDriver, err := backupstore.GetBackupStoreDriver(endpoint)
+	if err != nil {
+		return nil, err
+	}
+	return bsDriver, nil
+}
