+++ pkg/util/constants.go
@@ -111,4 +111,11 @@ const (
 	MaintainModeStrategyShutdownAndRestartAfterEnable  = "ShutdownAndRestartAfterEnable"
 	MaintainModeStrategyShutdownAndRestartAfterDisable = "ShutdownAndRestartAfterDisable"
 	MaintainModeStrategyShutdown                       = "Shutdown"
+
+	// s3 backup target constants
+	AWSAccessKey       = "AWS_ACCESS_KEY_ID"
+	AWSSecretKey       = "AWS_SECRET_ACCESS_KEY"
+	AWSEndpoints       = "AWS_ENDPOINTS"
+	AWSCERT            = "AWS_CERT"
+	VirtualHostedStyle = "VIRTUAL_HOSTED_STYLE"
 )
